+++ src/gitops/argocd.go
@@ -72,7 +72,7 @@ func buildArgoApplication(component string, artifact GitOpsArtifact, namespace s
 			},
 			"finalizers": []any{"resources-finalizer.argocd.argoproj.io"}, // ensure resources are deleted when app is deleted
 			"spec": map[string]any{
-				"project": "default",
+				"project": getArgoProject(artifact),
 				"source": map[string]any{
 					"repoURL":        artifact.HelmChart.Repository,
 					"chart":          artifact.HelmChart.Chart,
@@ -88,7 +88,7 @@ func buildArgoApplication(component string, artifact GitOpsArtifact, namespace s
 						"prune":    true,
 						"selfHeal": true,
 					},
-					"syncOptions": []any{"CreateNamespace=true"},
+					"syncOptions": []any{"CreateNamespace=true", "ServerSideApply=true"},
 				},
 			},
 		},
@@ -108,7 +108,7 @@ func buildArgoMoacApplication(component string, artifact GitOpsArtifact, namespa
 			},
 			"finalizers": []any{"resources-finalizer.argocd.argoproj.io"}, // ensure resources are deleted when app is deleted
 			"spec": map[string]any{
-				"project": "default",
+				"project": getArgoProject(artifact),
 				"source": map[string]any{
 					"repoURL":        moacRepository,
 					"chart":          moacChart,
@@ -129,9 +129,16 @@ func buildArgoMoacApplication(component string, artifact GitOpsArtifact, namespa
 						"prune":    true,
 						"selfHeal": true,
 					},
-					"syncOptions": []any{"CreateNamespace=true"},
+					"syncOptions": []any{"CreateNamespace=true", "ServerSideApply=true"},
 				},
 			},
 		},
 	}
+}
+
+func getArgoProject(artifact GitOpsArtifact) string {
+	if artifact.ArgoCD == nil || artifact.ArgoCD.Project == "" {
+		return "default"
+	}
+	return artifact.ArgoCD.Project
 }