+++ src/reconciler/platform_traefik.go
@@ -15,7 +15,7 @@ func (d *reconcilerModule) reconcileTraefik(ctx context.Context, spec v1alpha1.P
 		componentSpec{
 			enabled:          t.Enabled,
 			chart:            t.Chart,
-			patch:            t.Patch,
+			patches:          t.Patches,
 			name:             componentTraefik,
 			defaultChart:     "traefik",
 			defaultRepo:      "https://helm.traefik.io/traefik",
@@ -26,7 +26,13 @@ func (d *reconcilerModule) reconcileTraefik(ctx context.Context, spec v1alpha1.P
 			return []any{}, nil
 		},
 		func(ctx context.Context) (map[string]any, error) {
-			return nil, nil
+			values := map[string]any{}
+
+			if spec.Traefik.Service != nil {
+				values["service"] = spec.Traefik.Service
+			}
+
+			return values, nil
 		},
 	)
 }